Use Typesafe config for settings
https://github.com/typesafehub/config

Practically standard in the Scala world where application developers
can have library configs merged in with their app's.

Updated all the references in the request methods to use `settings.key`
instead of setting local vals for them in order to set the stage for
optional keys noted in a TODO -- they would become lazy vals and we'd
defer referencing them until we want to handle within a Try.